Added flatbuffers
Added flatbuffers as a dependency for message serialization.

We can create a custom schema for messaging types for communication between server/client.

The client manager now stores clients based on the ENetPeer which can be simply looked up and used an a unique identifier. This means we don't have to generate a random ID for each client.

NOTE: This is incomplete.